restructure player components
