+++ src/modules/player/subcomponents/comments.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import CommentsItem from './comments_item';
-import NewComment from './new_comment';
+import { CommentsItem } from './comments_item';
+import { NewComment } from './new_comment';
 
 class Comments extends React.Component {
   constructor(props) {
@@ -47,4 +47,4 @@ Comments.propTypes = {
   videoId: PropTypes.string
 };
 
-export default Comments;
+export { Comments };
