+++ src/modules/player/subcomponents/index.js
@@ -1,0 +1,10 @@
+export * from './comments_item';
+export * from './comments';
+export * from './details_lower';
+export * from './details_upper';
+export * from './details';
+export * from './like_dislike_ratio';
+export * from './new_comment';
+export * from './player';
+export * from './related_list_item';
+export * from './related';
