+++ src/modules/player/subcomponents/like_dislike_ratio.jsx
@@ -19,4 +19,4 @@ LikeDislikeRatio.propTypes = {
   ratio: PropTypes.number
 };
 
-export default LikeDislikeRatio;
+export { LikeDislikeRatio };
