+++ src/modules/player/subcomponents/player.jsx
@@ -39,4 +39,4 @@ Player.propTypes = {
   videoId: PropTypes.string
 };
 
-export default Player;
+export { Player };
